+++ tests/testthat/test-doc2vec.R
@@ -2,7 +2,8 @@ library(quanteda)
 library(wordvector)
 options(wordvector_threads = 2)
 
-corp <- head(data_corpus_inaugural, 59)
+corp <- head(data_corpus_inaugural, 59) %>% 
+    corpus_reshape()
 
 toks <- tokens(corp, remove_punct = TRUE, remove_symbols = TRUE,
                concatenator = " ") %>% 
@@ -22,13 +23,13 @@ test_that("textmodel_doc2vec works", {
         names(dov1),
         c("values", "weights", "type", "dim", "frequency", "window",  "iter", "alpha", 
           "use_ns", "ns_size", "sample", "normalize",  "min_count", "tolower",
-          "concatenator", "docvars", "call", "version")
+          "concatenator", "docvars", "ntoken", "call", "version")
     )
     expect_equal(
         dim(dov1$values$word), c(5363L, 50L)
     )
     expect_equal(
-        dim(dov1$values$doc), c(59L, 50L)
+        dim(dov1$values$doc), c(5296L, 50L)
     )
     expect_output(
         print(dov1),
@@ -38,7 +39,7 @@ test_that("textmodel_doc2vec works", {
             "textmodel_doc2vec(x = toks, dim = 50, min_count = 2, iter = 5, ",
             "    verbose = TRUE)",
             "",
-            "50 dimensions; 59 documents.", sep = "\n"), fixed = TRUE
+            "50 dimensions; 5,296 documents.", sep = "\n"), fixed = TRUE
     )
     expect_equal(
         class(expect_output(print(dov1))), 
@@ -59,6 +60,11 @@ test_that("textmodel_doc2vec works", {
         rownames(probability(dov1, c("good", "bad"), layer = "documents", mode = "numeric",
                              group = TRUE)),
         levels(dov1$docvars$docid_)
+    )
+    
+    expect_identical(
+        dov1$ntoken,
+        ntoken(tokens_trim(tokens_tolower(toks), min_termfreq = 2), remove_padding = TRUE)
     )
     
     expect_true(
@@ -85,13 +91,13 @@ test_that("textmodel_doc2vec works", {
         names(dov2),
         c("values", "weights", "type", "dim", "frequency", "window",  "iter", "alpha", 
           "use_ns", "ns_size", "sample", "normalize",  "min_count", "tolower",
-          "concatenator", "docvars", "call", "version")
+          "concatenator", "docvars", "ntoken", "call", "version")
     )
     expect_null(
         dov2$values$word
     )
     expect_equal(
-        dim(dov2$values$doc), c(59L, 50L)
+        dim(dov2$values$doc), c(5296L, 50L)
     )
     expect_output(
         print(dov2),
@@ -101,7 +107,7 @@ test_that("textmodel_doc2vec works", {
             "textmodel_doc2vec(x = toks, dim = 50, type = \"dbow\", min_count = 2, ", 
             "    iter = 5, verbose = TRUE)",
             "",
-            "50 dimensions; 59 documents.", sep = "\n"), fixed = TRUE
+            "50 dimensions; 5,296 documents.", sep = "\n"), fixed = TRUE
     )
     expect_equal(
         class(expect_output(print(dov2))), 
@@ -122,6 +128,11 @@ test_that("textmodel_doc2vec works", {
         rownames(probability(dov2, c("good", "bad"), layer = "documents", mode = "numeric",
                              group = TRUE)),
         levels(dov2$docvars$docid_)
+    )
+    
+    expect_identical(
+        dov2$ntoken,
+        ntoken(tokens_trim(tokens_tolower(toks), min_termfreq = 2), remove_padding = TRUE)
     )
     
     expect_true(